% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag.R
\docType{class}
\name{zigzag-class}
\alias{zigzag-class}
\alias{zigzag}
\title{zigzag reference class.}
\usage{
zigzag$new(data,
gene_length = NULL,
candidate_gene_list = "random",
num_active_components = 1,
weight_active_shape_1 = 2,
weight_active_shape_2 = 2,
inactive_means_prior_shape = 1,
inactive_means_prior_rate = 1/3,
inactive_variances_prior_min = 0.01,
inactive_variances_prior_max = 5,
spike_prior_shape_1 = 1,
spike_prior_shape_2 = 1,
active_means_dif_prior_shape = 1,
active_means_dif_prior_rate = 1/3,
active_variances_prior_min = 0.01,
active_variances_prior_max = 5,
shared_active_variances = TRUE,
output_directory = "output",
multi_ta = FALSE,
threshold_a = 0,
threshold_i = threshold_a[1],
beta = 1,
temperature = 1,
tau_shape = 1,
tau_rate = 1,
s0_mu = -1,
s0_sigma = 2,
s1_shape = 1,
s1_rate = 2,
alpha_r_shape = 1,
alpha_r_rate = 1/10,
active_gene_set = NULL, ...)#'
}
\description{
zigzag reference class.
}
\section{Fields}{

\describe{
\item{\code{gen}}{.}

\item{\code{field_list_values}}{.}

\item{\code{field_list_names}}{.}

\item{\code{sqrt2pi}}{.}

\item{\code{active_gene_set}}{.}

\item{\code{active_gene_set_idx}}{.}

\item{\code{Xg}}{.}

\item{\code{Yg}}{.}

\item{\code{Yg_proposed}}{.}

\item{\code{Yg_trace}}{.}

\item{\code{output_directory}}{.}

\item{\code{gene_names}}{.}

\item{\code{num_libraries}}{.}

\item{\code{num_active_components}}{.}

\item{\code{component_matrix}}{.}

\item{\code{num_transcripts}}{.}

\item{\code{gene_lengths}}{.}

\item{\code{multi_thresh_a}}{.}

\item{\code{threshold_a}}{.}

\item{\code{threshold_i}}{.}

\item{\code{inf_tol}}{.}

\item{\code{model_paramlist}}{.}

\item{\code{lnl_trace}}{.}

\item{\code{inactive_mean_tuningParam}}{.}

\item{\code{inactive_variance_tuningParam}}{.}

\item{\code{spike_probability_tuningParam}}{.}

\item{\code{active_mean_tuningParam}}{.}

\item{\code{active_variance_tuningParam}}{.}

\item{\code{mixture_weight_tuningParam}}{.}

\item{\code{tuningParam_s0}}{.}

\item{\code{tuningParam_s1}}{.}

\item{\code{tuningParam_tau}}{.}

\item{\code{tuningParam_s0tau}}{.}

\item{\code{tuningParam_alpha_r}}{.}

\item{\code{tuningParam_yg}}{.}

\item{\code{tuningParam_sigma_g}}{.}

\item{\code{tuningParam_multi_sigma}}{.}

\item{\code{tuningParam_sigma_mu}}{.}

\item{\code{multi_sigma_trace}}{.}

\item{\code{sigma_mu_trace}}{.}

\item{\code{active_idx}}{.}

\item{\code{inactive_idx}}{.}

\item{\code{beta}}{.}

\item{\code{beta_oneLib}}{.}

\item{\code{temperature}}{.}

\item{\code{proposal_vector}}{.}

\item{\code{proposal_probs}}{.}

\item{\code{level_1_probs}}{.}

\item{\code{hieararchical_probs_oneLibrary}}{.}

\item{\code{combined_probs}}{.}

\item{\code{weight_active_shape_1}}{.}

\item{\code{weight_active_shape_2}}{.}

\item{\code{inactive_means_prior_shape}}{.}

\item{\code{inactive_means_prior_rate}}{.}

\item{\code{inactive_variances_prior_min}}{.}

\item{\code{inactive_variances_prior_max}}{.}

\item{\code{inactive_variances_prior_shape}}{.}

\item{\code{inactive_variances_prior_rate}}{.}

\item{\code{inactive_variances_prior_log_min}}{.}

\item{\code{inactive_variances_prior_log_max}}{.}

\item{\code{spike_prior_shape_1}}{.}

\item{\code{spike_prior_shape_2}}{.}

\item{\code{active_means_dif_prior_shape}}{.}

\item{\code{active_means_dif_prior_rate}}{.}

\item{\code{active_variances_prior_min}}{.}

\item{\code{active_variances_prior_max}}{.}

\item{\code{active_variances_prior_shape}}{.}

\item{\code{active_variances_prior_rate}}{.}

\item{\code{active_variances_prior_log_min}}{.}

\item{\code{active_variances_prior_log_max}}{.}

\item{\code{weight_within_active_alpha}}{.}

\item{\code{s0_mu}}{.}

\item{\code{s0_sigma}}{.}

\item{\code{s1_rate}}{.}

\item{\code{s1_shape}}{.}

\item{\code{Sg}}{.}

\item{\code{s0}}{.}

\item{\code{s0_trace}}{.}

\item{\code{s1}}{.}

\item{\code{s1_trace}}{.}

\item{\code{tau_shape}}{.}

\item{\code{tau_rate}}{.}

\item{\code{tau}}{.}

\item{\code{tau_trace}}{.}

\item{\code{sigma_g}}{.}

\item{\code{sigma_g_trace}}{.}

\item{\code{s0tau_trace}}{.}

\item{\code{alpha_r_shape}}{.}

\item{\code{alpha_r_rate}}{.}

\item{\code{p_x}}{.}

\item{\code{alpha_r}}{.}

\item{\code{alpha_r_trace}}{.}

\item{\code{all_means}}{.}

\item{\code{all_variances}}{.}

\item{\code{no_detect_spike}}{.}

\item{\code{inactive_spike_allocation}}{.}

\item{\code{in_spike_idx}}{.}

\item{\code{out_spike_idx}}{.}

\item{\code{all_zero_idx}}{.}

\item{\code{candidate_gene_list}}{.}

\item{\code{XgLikelihood}}{.}

\item{\code{YgLikelihood}}{.}

\item{\code{sigma_g_probability}}{.}

\item{\code{allocation_active_inactive}}{.}

\item{\code{allocation_active_inactive_proposed}}{.}

\item{\code{allocation_trace}}{.}

\item{\code{allocation_active_inactive_prob}}{.}

\item{\code{allocation_active_inactive_prob_proposed}}{.}

\item{\code{inactive_means}}{.}

\item{\code{inactive_means_proposed}}{.}

\item{\code{inactive_means_trace}}{.}

\item{\code{inactive_means_prob}}{.}

\item{\code{inactive_means_prob_proposed}}{.}

\item{\code{inactive_variances}}{.}

\item{\code{inactive_variances_proposed}}{.}

\item{\code{inactive_variances_trace}}{.}

\item{\code{inactive_variances_prob}}{.}

\item{\code{inactive_variances_prob_proposed}}{.}

\item{\code{spike_probability}}{.}

\item{\code{spike_probability_proposed}}{.}

\item{\code{spike_probability_trace}}{.}

\item{\code{spike_probability_prob}}{.}

\item{\code{spike_probability_prob_proposed}}{.}

\item{\code{active_means_dif}}{.}

\item{\code{active_means_dif_proposed}}{.}

\item{\code{active_means_trace}}{.}

\item{\code{active_means_dif_prob}}{.}

\item{\code{active_means_dif_prob_proposed}}{.}

\item{\code{active_means}}{.}

\item{\code{shared_active_variances}}{.}

\item{\code{active_variances}}{.}

\item{\code{active_variances_proposed}}{.}

\item{\code{active_variances_trace}}{.}

\item{\code{active_variances_prob}}{.}

\item{\code{active_variances_prob_proposed}}{.}

\item{\code{weight_active}}{.}

\item{\code{weight_active_proposed}}{.}

\item{\code{weight_active_trace}}{.}

\item{\code{weight_active_prob}}{.}

\item{\code{weight_active_prob_proposed}}{.}

\item{\code{weight_within_active}}{.}

\item{\code{weight_within_active_proposed}}{.}

\item{\code{weight_within_active_trace}}{.}

\item{\code{weight_within_active_prob}}{.}

\item{\code{weight_within_active_prob_proposed}}{.}

\item{\code{mixture_weight_trace}}{.}

\item{\code{allocation_within_active}}{.}

\item{\code{allocation_within_active_proposed}}{.}
}}


