% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzagMCMC.R
\name{mcmc}
\alias{mcmc}
\title{mcmc}
\usage{
my_zigzag_object$mcmc(sample_frequency = 10, ngen = 10000, target_ESS = NULL,
progress_plot = FALSE, write_to_files = TRUE, mcmcprefix = "out",
compute_probs = TRUE, run_posterior_predictive = TRUE, append = FALSE)
}
\arguments{
\item{sample_frequency}{Number of generations between samples from the chain}

\item{ngen}{Number of generations to run the chain}

\item{target_ESS}{Run chain until the target effective sample size is achieved for the likelihood}

\item{progress_plot}{Display a plot showing summary statistics and graphs. (in Rstudio)}

\item{write_to_files}{Write posterior samples to output files located in mcmcprefix_mcmc_ouput directory}

\item{mcmcprefix}{prefix for mcmc output directory as well as the output files within.}

\item{compute_probs}{compute probabilities of activity for each gene. Default is TRUE.}

\item{run_posterior_predictive}{run posterior predictive simulation for every other sample from the chain.}

\item{append}{If posterior sample log files in mcmcprefix_mcmc_ouput already exists, then append samples to those files. Default is FALSE.}
}
\value{
blorb
}
\description{
MCMC sampling from hieararchical bayesian mixture model
}
\examples{
{
\dontrun{mcmc(sample_frequency = 100, ngen = 50000, mcmcprefix = "organ_x_posterior_sample")}
}

}
