% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag.R
\docType{class}
\name{zigzag-class}
\alias{zigzag-class}
\alias{zigzag}
\title{zigzag reference class.}
\usage{
zigzag$new(data, gene_length = NULL, candidate_gene_list = "random",
num_active_components = 1,
weight_active_shape_1 = 2,
weight_active_shape_2 = 2,
inactive_means_prior_shape = 1,
inactive_means_prior_rate = 1/3,
inactive_variances_prior_min = 0.01,
inactive_variances_prior_max = 5,
spike_prior_shape_1 = 1,
spike_prior_shape_2 = 1,
active_means_dif_prior_shape = 1,
active_means_dif_prior_rate = 1/3,
active_variances_prior_min = 0.01,
active_variances_prior_max = 5,
shared_active_variances = TRUE,
output_directory = "output",
threshold_a = 0,
threshold_i = threshold_a[1],
variance_g_upper_bound = Inf,
beta = 1,
tau_shape = 1,
tau_rate = 1,
s0_mu = -1,
s0_sigma = 2,
s1_shape = 1,
s1_rate = 2,
alpha_r_shape = 1,
alpha_r_rate = 1/10, ...)
}
\description{
This class contains all the tools for computing the posterior probability
each gene in a dataset is active given a set of libraries. Creating a zigzag object sets all the
hyperparameters and initializes all model parameters.
}
\section{Fields}{

\describe{
\item{\code{Xg}}{A G x R matrix containing the log-relative-expression of G genes in R libraries.}

\item{\code{Yg}}{A vector of length G containing the true latent log-relative-expression for G, genes.}

\item{\code{output_directory}}{A directory where burnin and and mcmc log files and plots are created and stored.}

\item{\code{gene_names}}{A character vector of length G contianing the gene names.}

\item{\code{num_libraries}}{The number of libraries in the data.}

\item{\code{num_active_components}}{The number of Normally-distributed subcomponents of the active distribution.}

\item{\code{num_transcripts}}{The number of genes or transcripts in the data depending on whether the data is gene or transcript level expression.}

\item{\code{gene_lengths}}{A vector of length G, containing either the mean lengths of gene transcripts or the lengths of transcripts.}

\item{\code{threshold_a}}{A vector containing the lower-bound of the prior distribution for the active component means.
If the length of the threshold_a vector is < than num_active_components then threshold_a[1] is used as the lower bound for the first active component
and subsequent active means are parameterized with orderd offsets rather than component-specific
lower boundaries. For example, if the user specifies 3 active subcomponents and sets threshold_a = c(1,3), then
the lower boundary for subcomponent 1 will be set 1, and the prior distributions for the difference between mean 1 and 2,
and the difference between mean 2 and 3 will be gamma distributed.}

\item{\code{threshold_i}}{The upper boundary for the inactive distribution.}

\item{\code{candidate_gene_list}}{A character vector of genes of interest for which mcmc log files for Yg and variance_g should be output.}

\item{\code{shared_active_variances}}{A boolean indicating if all active subcomponents share a variance component. If FALSE, a variance parameter is estimated for each component.}

\item{\code{variance_g_upper_bound}}{The upper boundary for the truncated log-normal prior for gene-specific variances. Default = Inf}

\item{\code{weight_active_shape_1}}{.}

\item{\code{weight_active_shape_2}}{.}

\item{\code{inactive_means_prior_shape}}{.}

\item{\code{inactive_means_prior_rate}}{.}

\item{\code{inactive_variances_prior_min}}{.}

\item{\code{inactive_variances_prior_max}}{.}

\item{\code{inactive_variances_prior_shape}}{.}

\item{\code{inactive_variances_prior_rate}}{.}

\item{\code{inactive_variances_prior_log_min}}{.}

\item{\code{inactive_variances_prior_log_max}}{.}

\item{\code{spike_prior_shape_1}}{.}

\item{\code{spike_prior_shape_2}}{.}

\item{\code{active_means_dif_prior_shape}}{.}

\item{\code{active_means_dif_prior_rate}}{.}

\item{\code{active_variances_prior_min}}{.}

\item{\code{active_variances_prior_max}}{.}

\item{\code{active_variances_prior_shape}}{.}

\item{\code{active_variances_prior_rate}}{.}

\item{\code{active_variances_prior_log_min}}{.}

\item{\code{active_variances_prior_log_max}}{.}

\item{\code{weight_within_active_alpha}}{.}

\item{\code{s0_mu}}{.}

\item{\code{s0_sigma}}{.}

\item{\code{s1_rate}}{.}

\item{\code{s1_shape}}{.}

\item{\code{Sg}}{.}

\item{\code{s0}}{.}

\item{\code{s1}}{.}

\item{\code{tau_shape}}{.}

\item{\code{tau_rate}}{.}

\item{\code{tau}}{.}

\item{\code{variance_g}}{.}

\item{\code{s0tau_trace}}{.}

\item{\code{alpha_r_shape}}{.}

\item{\code{alpha_r_rate}}{.}

\item{\code{p_x}}{.}

\item{\code{alpha_r}}{.}

\item{\code{inactive_spike_allocation}}{.}

\item{\code{candidate_gene_list}}{.}

\item{\code{allocation_active_inactive}}{.}

\item{\code{allocation_active_inactive_prob}}{.}

\item{\code{inactive_means}}{.}

\item{\code{inactive_variances}}{.}

\item{\code{spike_probability}}{.}

\item{\code{active_means_dif}}{.}

\item{\code{active_means}}{.}

\item{\code{active_variances}}{.}

\item{\code{weight_active}}{.}

\item{\code{weight_within_active}}{.}

\item{\code{allocation_within_active}}{.}

\item{\code{beta}}{A power exponent for the likelihood function. If set to 0, then the mcmc samples from the joint prior.}
}}


